Improve upload with plupload
plupload is a uploader and offers html5, flash, and silverlight modules to
support multiple file, upload progress and chunked uploads with drag and drop
funtionaliaty.

With plupload, phtagr offers a chunked upload of files for supported internet
browsers. Huge files like HD movies could be uploaded without changing php
settings for large POST data. The progress of uploaded files are shown as well.

The integration of plupload is done via backbone to play around with this gread
js framework.